+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2026-03-27
+
 - **BREAKING**: Config path moved from `~/.dirty-repo-scanner.yml` to `~/.config/dirty-repo-scanner/config.yml` (XDG compliant). Move your config file manually.
 - Drop `go-homedir` dependency in favor of `os.UserConfigDir()`
 - Detect running in tmux and then run e tmux popup